Change build to web pack and update to `ES2015`
This happens because of many reasons.
First of all `I like webpack` ;).
Besides this it relieved the dev, build and test
processes. Some code changes happened but most of
the changes resulting from folder renaming and
deleting of obsolete files.
Also the test frame work change from `jasmine` to
`mocha` because of the flexibility of the assert
and mock library.

* removed bower and all related files
* add webpack config file
* add node modules
* update README file
* change template file extension to `ejs`
* remove card data for dev use (data_bak)
* merge karma configs into one file
* and many more